// src/promotions/enums/promotion-type.enum.ts
export enum PromotionType {
  NXN = 'NXN', // Compra N, llévate N gratis
  PERCENT_SECOND = 'PERCENT_SECOND', // Descuento porcentual en el segundo producto
  PERCENTAGE = 'PERCENTAGE', // Descuento porcentual en el producto
  FIXED = 'FIXED', // Descuento fijo en el producto
  BUNDLE = 'BUNDLE', // Paquete de productos con descuento
}